+++ src/flamegraph.rs
@@ -195,11 +195,11 @@ pub struct FlamegraphArgs {
         help_heading = "Linker"
     )]
     pub search_path: Vec<PathBuf>,
-    /// Link compiled projects to the specified library NAME.
+    /// Load the compiled library package NAME.
     #[arg(
         long = "link-library",
         short = 'l',
-        value_name = "[KIND=]NAME",
+        value_name = "[masp[:LINKAGE]=]NAME",
         value_delimiter = ',',
         next_line_help(true),
         help_heading = "Linker"